+++ common/keyboard.nix
@@ -5,11 +5,12 @@
 ##                                                                            ##
 ################################################################################
 
-{ config, lib, pkgs, ... }:
+{ lib, ... }:
 
 let
-  inherit (lib) mkOption types;
-  layouts = types.enum [ null "bépo" ];
+  inherit (lib) mkOption;
+  inherit (lib.types) enum;
+  layouts = enum [ null "bépo" ];
 in
 
 {