+++ nixos/modules/programs/tmux.nix
@@ -15,15 +15,15 @@
 
 let
   inherit (builtins) readFile;
-  inherit (lib) mkOption mkEnableOption mkIf mkDefault types;
-  inherit (lib.trivial) release;
+  inherit (lib) mkDefault mkEnableOption mkIf mkOption types;
   inherit (pkgs) stdenv;
 
   cfg = config.confkit.programs.tmux;
 
-  tmuxConfig = if cfg.bepo
-               then readFile ../../../misc/tmux_bepo.conf
-               else readFile ../../../misc/tmux.conf;
+  tmuxConfig =
+    if cfg.bepo
+    then readFile ../../../misc/tmux_bepo.conf
+    else readFile ../../../misc/tmux.conf;
 in
 
 {
@@ -45,12 +45,12 @@ in
       enable = true;
     } // (if stdenv.isDarwin then {
       tmuxConfig = mkDefault (tmuxConfig + ''
-          bind-key -T copy-mode Enter send-keys -X copy-pipe-and-cancel "${pkgs.reattach-to-user-namespace}/bin/reattach-to-user-namespace pbcopy"
-        '');
+        bind-key -T copy-mode Enter send-keys -X copy-pipe-and-cancel "${pkgs.reattach-to-user-namespace}/bin/reattach-to-user-namespace pbcopy"
+      '');
     } else {
       extraConfig = mkDefault (tmuxConfig + ''
-          bind-key -T copy-mode Enter send-keys -X copy-selection-and-cancel
-        '');
+        bind-key -T copy-mode Enter send-keys -X copy-selection-and-cancel
+      '');
     });
   };
 }